Update to C++ API

+++ source/OpenCL_matrixfree_multi_gpu.cpp
@@ -190,7 +190,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray*prhs[]) {
 		ind++;
 
 		// The total number of sinograms
-		const uint16_t TotSinos = (uint16_t)mxGetScalar(prhs[ind]);
+		const uint32_t TotSinos = (uint32_t)mxGetScalar(prhs[ind]);
 		ind++;
 		
 		uint32_t loop_var_par = 1u;
