+++ source/multidevice_kernel.cl
@@ -136,7 +136,7 @@ void siddon_precomp(const uint d_Nxy, const uint d_N, const uint d_Nx, const uin
 
 void kernel_multi(const float global_factor, const float d_epps, const uint d_N, const uint d_Nx, const uint d_Ny, const uint d_Nz, 
 	const float d_dz, const float d_dx,	const float d_dy, const float d_bz, const float d_bx, const float d_by, const float d_bzb, const float d_maxxx, 
-	const float d_maxyy, const float d_zmax, const float d_NSlices, const uint d_size_x, const ushort d_TotSinos, 
+	const float d_maxyy, const float d_zmax, const float d_NSlices, const uint d_size_x, const uint d_TotSinos, 
 	const uint d_det_per_ring, const uint d_pRows, const uint d_Nxy, const uchar fp,
 	const float tube_width_xy, const float crystal_size_z, const float bmin, const float bmax, const float Vmax, const float d_epsilon_mramla, 
 	const __global float* d_atten, __constant uint* d_pseudos, const __global float* d_x, const __global float* d_y, const __global float* d_zdet,
